# docs/api/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Artistic Hub API",
        "version": "0.0.1",
        "description": "Artistic Hub API: public catalog, customer accounts, cart, checkout and orders, and the admin API under /admin."
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1"
        }
    ],
    "security": [
        {
            "http": []
        }
    ],
    "paths": {
...
        "/admin/products/{product}": {
            "get": {
                "operationId": "admin.v1.products.show",
                "summary": "View a product with its variants, photos and tags",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product public id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "`AdminProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminProductResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "variants"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "put": {
                "operationId": "admin.v1.products.update",
                "description": "Send the whole product: variants with a `public_id` are updated, without one are created, and variants not\nsent are deleted. Turning the product off turns every variant off. Photos not sent are deleted with their files.",
                "summary": "Save a product with all its variants",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product public id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`AdminProductResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/AdminProductResource"
                                                },
                                                {
                                                    "type": "object",
                                                    "required": [
                                                        "variants"
                                                    ]
                                                }
                                            ]
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.v1.products.destroy",
                "description": "Deletes its variants, photos (and their files) and tag links. Past orders keep their own copy.",
                "summary": "Delete a product",
                "tags": [
                    "Product"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product public id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
        "/admin/products/{product}/variants/{variant}": {
            "delete": {
                "operationId": "admin.v1.products.variants.destroy",
                "description": "Removes it right away (outside the product form), with its photos and tag links. A product must keep at\nleast one variant (409). An id that is not this product's variant returns 404.",
                "summary": "Delete one variant",
                "tags": [
                    "ProductVariant"
                ],
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "required": true,
                        "description": "The product public id",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "variant",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/customer/profile/avatar": {
            "post": {
                "operationId": "customer.v1.profile.avatar.update",
                "description": "multipart/form-data with an `avatar` file. Replaces any previous avatar. Returns the profile with the new `avatar_url`.",
                "summary": "Upload or replace the avatar",
                "tags": [
                    "ProfileAvatar"
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateAvatarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`CustomerProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerProfileResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    }
                }
            },
            "delete": {
                "operationId": "customer.v1.profile.avatar.destroy",
                "summary": "Remove the avatar",
                "tags": [
                    "ProfileAvatar"
                ],
                "responses": {
                    "200": {
                        "description": "`CustomerProfileResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/CustomerProfileResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "description": "Error overview."
                                        }
                                    },
                                    "required": [
                                        "message"
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/tags/{tag}": {
            "put": {
                "operationId": "admin.v1.tags.update",
                "summary": "Rename a tag (everywhere it is used)",
                "tags": [
                    "Tag"
                ],
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "description": "The tag public id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SaveTagRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "`TagResource`",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/TagResource"
                                        }
                                    },
                                    "required": [
                                        "data"
                                    ]
                                }
                            }
                        }
                    },
                    "422": {
                        "$ref": "#/components/responses/ValidationException"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            },
            "delete": {
                "operationId": "admin.v1.tags.destroy",
                "summary": "Delete a tag. It is removed from every variant; the variants stay",
                "tags": [
                    "Tag"
                ],
                "parameters": [
                    {
                        "name": "tag",
                        "in": "path",
                        "required": true,
                        "description": "The tag public id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        },
...
        "/admin/uploads/variant-photos/{media}": {
            "delete": {
                "operationId": "admin.v1.uploads.variant-photos.destroy",
                "description": "Deletes the photo and its file right away: one on a saved variant (its other photos keep their order), or\none uploaded and removed from the form before saving. Any admin can delete any variant photo; an id that is\nnot a variant photo returns 404.",
                "summary": "Delete a variant photo",
                "tags": [
                    "VariantPhoto"
                ],
                "parameters": [
                    {
                        "name": "media",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "$ref": "#/components/responses/ModelNotFoundException"
                    },
                    "401": {
                        "$ref": "#/components/responses/AuthenticationException"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "http": {
                "type": "http",
                "scheme": "bearer"
            }
        },
        "schemas": {
...
            "AdminProductResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Internal database id: shown for reference only; URLs and requests take public_id."
                    },
                    "public_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string"
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AdminProductVariantResource"
                        }
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "id",
                    "public_id",
                    "name",
                    "slug",
                    "description",
                    "is_active",
                    "created_at",
                    "updated_at"
                ],
                "title": "AdminProductResource"
            },
...
            "CustomerProfileResource": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Stored on the user account; set and changed with the profile."
                    },
                    "phone": {
                        "type": "string"
                    },
                    "date_of_birth": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "description": "Y-m-d, or null."
                    },
                    "gender": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Gender"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "avatar_url": {
                        "type": "string",
                        "description": "Public URL of the avatar, or null if none."
                    },
                    "created_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    },
                    "updated_at": {
                        "type": [
                            "string",
                            "null"
                        ]
                    }
                },
                "required": [
                    "name",
                    "phone",
                    "date_of_birth",
                    "gender",
                    "avatar_url",
                    "created_at",
                    "updated_at"
                ],
                "title": "CustomerProfileResource"
            },
...
            "SaveProductRequest": {
                "type": "object",
                "description": "The whole product, sent on every save (create and update): product fields plus every variant it\nshould have. Variants with a `public_id` are updated, without one are created, and variants not sent are\ndeleted. Photos are media ids returned by POST /admin/uploads/variant-photos, in display order.\nEvery id here is a client-facing public id (ULID); toData() turns them into internal ids.",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Unique across products, ignoring case.",
                        "maxLength": 255
                    },
                    "description": {
                        "type": [
                            "string",
                            "null"
                        ],
                        "maxLength": 5000
                    },
                    "is_active": {
                        "type": "boolean",
                        "description": "Turning the product off turns every variant off."
                    },
                    "variants": {
                        "type": "array",
                        "items": {
                            "type": "object",
                            "properties": {
                                "public_id": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "description": "The variant's public_id; leave out for a new variant."
                                },
                                "name": {
                                    "type": "string",
                                    "description": "Unique within this product, ignoring case (other products may use the same name).",
                                    "maxLength": 255
                                },
                                "sku": {
                                    "type": "string",
                                    "pattern": "^[A-Za-z0-9_-]+$",
                                    "maxLength": 64
                                },
                                "description": {
                                    "type": [
                                        "string",
                                        "null"
                                    ],
                                    "description": "Falls back to the product description when null.",
                                    "maxLength": 5000
                                },
                                "original_price": {
                                    "type": "string",
                                    "description": "Rupees, up to 2 decimals, e.g. \"999.00\".",
                                    "pattern": "^\\d{1,8}(\\.\\d{1,2})?$"
                                },
                                "selling_price": {
                                    "type": "string",
                                    "description": "Rupees, up to 2 decimals; must not be more than original_price.",
                                    "pattern": "^\\d{1,8}(\\.\\d{1,2})?$"
                                },
                                "stock": {
                                    "type": "integer",
                                    "minimum": 0,
                                    "maximum": 1000000
                                },
                                "is_active": {
                                    "type": "boolean"
                                },
                                "tag_ids": {
                                    "type": [
                                        "array",
                                        "null"
                                    ],
                                    "items": {
                                        "type": "string"
                                    },
                                    "uniqueItems": true
                                },
                                "photo_ids": {
                                    "type": [
                                        "array",
                                        "null"
                                    ],
                                    "description": "Up to 8 media ids per variant (from POST /admin/uploads/variant-photos), in display order; the first is the cover.",
                                    "items": {
                                        "type": "string"
                                    },
                                    "maxItems": 8,
                                    "uniqueItems": true
                                }
                            },
                            "required": [
                                "name",
                                "sku",
                                "original_price",
                                "selling_price",
                                "stock",
                                "is_active"
                            ]
                        },
                        "minItems": 1,
                        "maxItems": 50,
                        "uniqueItems": true
                    }
                },
                "required": [
                    "name",
                    "is_active",
                    "variants"
                ],
                "title": "SaveProductRequest"
            },
            "SaveTagRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string",
                        "description": "Unique; the slug is generated from it.",
                        "maxLength": 100
                    }
                },
                "required": [
                    "name"
                ],
                "title": "SaveTagRequest"
            },
...
            "TagResource": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "string",
                        "description": "Internal database id: shown for reference only; URLs and requests take public_id."
                    },
                    "public_id": {
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "public_id",
                    "name",
                    "slug"
                ],
                "title": "TagResource"
            },
...
            "UpdateAvatarRequest": {
                "type": "object",
                "properties": {
                    "avatar": {
                        "type": "string",
                        "format": "binary",
                        "contentMediaType": "application/octet-stream",
                        "description": "Maximum file size: 2048 kilobytes."
                    }
                },
                "required": [
                    "avatar"
                ],
                "title": "UpdateAvatarRequest"
            },
...
        },
        "responses": {
            "ValidationException": {
                "description": "Validation error",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Errors overview."
                                },
                                "errors": {
                                    "type": "object",
                                    "description": "A detailed description of each field that failed validation.",
                                    "additionalProperties": {
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            },
                            "required": [
                                "message",
                                "errors"
                            ]
                        }
                    }
                }
            },
...
            "AuthenticationException": {
                "description": "Unauthenticated",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            },
            "ModelNotFoundException": {
                "description": "Not found",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "object",
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "description": "Error overview."
                                }
                            },
                            "required": [
                                "message"
                            ]
                        }
                    }
                }
            }
        }
    }
}